Add `--memory` option to pytest
This prints the total amount of memory consumed by the Python
process running the unit tests after all of the test have completed.
It seems pretty consistent across runs so far.

Note that this only works on Linux and Mac OS X. Making this work
on Windows is possible using `win32api`, but it's not clear which
versions of Windows it works on. We could also use an external
library to do this (e.g., `psutils`) if the consensus demands it.

Also fix coverage.sh template to allow multiple arguments.

Co-authored-by: Eric Hunsberger <eric.hunsberger@appliedbrainresearch.com>