squash! Nengo-wide default dtype can be set in RC
The main change is to simplify the dtype implementation.
Previously, the idea was to do this through the RC system
to set the default value, but then you could change it
on the fly through arguments to the simulator or model.
This gives two ways to do something, which I could understand
if one way was particularly onerous, but they're pretty much
the same, so might as well give only one way to do it. Also
cuts down on the surface area of the API by minimizing arguments.

Since the behavior of the RC setting is slightly different from
the argument to the simulator, I made a second RC setting that
defaults to the value of the first one. This is really only
for the very specific case that you would want them to be
different, and honestly I think we can probably lose the second
one, but I didn't want to change behavior in case that was an
important thing for whatever reason.

Other changes:

- Switch order of --memory and --plots, otherwise the plot
  directory will be called '--memory'
- Remove 'global_dtype' variable in favor of a property on
  `nengo.rc`. The 'global_dtype' variable won't update when
  the RC setting changes, and it ends up being shorter to
  access it on the rc object anyway.
- Fixed Sparse transform docstring, the indices entry was
  incomplete